♻️ refactor(SCRUM-210): PortfolioInfo interface model 폴더에 분리
