+++ src/model/index.ts
@@ -1,0 +1,9 @@
+import { Job } from '@/model/job';
+
+export interface PortfolioInfo {
+  title: string;
+  author: string;
+  job: Job;
+  likes: number;
+  createdAt: string;
+}
